Rename `test` directory to `tests`
