+++ web-local/tests/dashboards/number-formatting.spec.ts
