+++ web-local/tests/dashboards/time-controls-from-config.spec.ts
@@ -1,11 +1,11 @@
 import { expect, test } from "@playwright/test";
-import { useDashboardFlowTestSetup } from "web-local/test/dashboards/dashboard-flow-test-setup";
-import { updateCodeEditor } from "web-local/test/utils/commonHelpers";
+import { useDashboardFlowTestSetup } from "web-local/tests/dashboards/dashboard-flow-test-setup";
+import { updateCodeEditor } from "web-local/tests/utils/commonHelpers";
 import {
   interactWithTimeRangeMenu,
   waitForDashboard,
-} from "web-local/test/utils/dashboardHelpers";
-import { startRuntimeForEachTest } from "web-local/test/utils/startRuntimeForEachTest";
+} from "web-local/tests/utils/dashboardHelpers";
+import { startRuntimeForEachTest } from "web-local/tests/utils/startRuntimeForEachTest";
 
 test.describe("time controls settings from dashboard config", () => {
   startRuntimeForEachTest();
