+++ web-local/tests/utils/sourceHelpers.ts
@@ -13,7 +13,7 @@ import { waitForEntity } from "./waitHelpers";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
-export const TestDataPath = path.join(__dirname, "../../data");
+export const TestDataPath = path.join(__dirname, "../data");
 
 /**
  * Used to upload local file as a source
